Address exact-state scheduler review feedback
